Move `<MultilineTextFieldEdit>` into field editors
