+++ designer/client/src/ComponentTypeEdit.tsx
@@ -7,13 +7,13 @@ import React, {
 } from 'react'
 
 import { FieldEdit } from '~/src/FieldEdit.jsx'
-import { MultilineTextFieldEdit } from '~/src/MultilineTextFieldEdit.jsx'
 import { ConditionEdit } from '~/src/components/FieldEditors/ConditionEdit.jsx'
 import { ContentEdit } from '~/src/components/FieldEditors/ContentEdit.jsx'
 import { DateFieldEdit } from '~/src/components/FieldEditors/DateFieldEdit.jsx'
 import { FileUploadFieldEdit } from '~/src/components/FieldEditors/FileUploadFieldEdit.jsx'
 import { ListContentEdit } from '~/src/components/FieldEditors/ListContentEdit.jsx'
 import { ListFieldEdit } from '~/src/components/FieldEditors/ListFieldEdit.jsx'
+import { MultilineTextFieldEdit } from '~/src/components/FieldEditors/MultilineTextFieldEdit.jsx'
 import { NumberFieldEdit } from '~/src/components/FieldEditors/NumberFieldEdit.jsx'
 import { SelectFieldEdit } from '~/src/components/FieldEditors/SelectFieldEdit.jsx'
 import { TextFieldEdit } from '~/src/components/FieldEditors/TextFieldEdit.jsx'
